new passive `MediaRecorder` system
```
With the new MediaRecorder system. We have two recorders. Both are recording
for 2 seconds, but offset by 1s each. So if at 1500ms we detect that a
handstand has been happening for the last 1s, we set activeMediaRecorder
to point to passiveMediaRecorder1, and stop passiveMediaRecorder2, as there
would be 1500ms of footage available on passiveMediaRecorder1, but only
500ms of footage available on passiveMediaRecorder2. (passiveMediaRecorder2
started recording 1000ms after passiveMediaRecorder1.)
                   ▾
      0s       1s       2s       3s      4s       5s
  1   |█████████████░░░░|░░░░░░░░░░░░░░░░░|░░░░░░░░░ ->
  2            |▒▒▒▒░░░░░░░░░░░░░|░░░░░░░░░░░░░░░░░|

If we were to detect it at 1800ms, we would still use passiveMediaRecorder1,
as there were 1800ms of footage available there, but only 800ms of footage
available on passiveMediaRecorder2.
                      ▾
      0s       1s       2s       3s      4s       5s
  1   |████████████████░|░░░░░░░░░░░░░░░░░|░░░░░░░░░ ->
  2            |▒▒▒▒▒▒▒░░░░░░░░░░|░░░░░░░░░░░░░░░░░|

However, at 2100ms, we would instead use passiveMediaRecorder2, as there
are 1100ms of footage available there. At 2000ms, passiveMediaRecorder1
was stopped and the recording was deleted, and then restarted.
                         ▾
      0s       1s       2s       3s      4s       5s
  1   |░░░░░░░░░░░░░░░░░|▒░░░░░░░░░░░░░░░░|░░░░░░░░░ ->
  2            |█████████████████|░░░░░░░░░░░░░░░░░|
```